+++ packages/fonts/src/types/payload-augment.ts
@@ -4,4 +4,6 @@
 import type { DefaultRecord, Get } from '@pro-laico/core'
 
 export type Font = Get<'Font', DefaultRecord>
+export type FontFile = Get<'FontFile', DefaultRecord>
 export type FontSet = Get<'FontSet', DefaultRecord>
+export type FontOriginal = Get<'FontOriginal', DefaultRecord>