+++ templates/atomic-payload/src/payload-types.augment.d.ts
@@ -10,6 +10,8 @@ declare module '@pro-laico/core' {
     CollectionWithStoredAtomicClassesSlug: G.CollectionWithStoredAtomicClassesSlug
     DesignSet: G.DesignSet
     Font: G.Font
+    FontFile: G.FontFile
+    FontOriginal: G.FontOriginal
     Footer: G.Footer
     GeneratedImage: G.GeneratedImage
     Header: G.Header